MINOR: zrancid-ssh: Add sshchain via mode
I came across a case where ssh -J (ProxyJump) cannot connect to a device.
This was random and happened on el8 but not on el9.

[zrancid@jth-rocky8 ~]$ 'ssh' '-o' 'ConnectTimeout=10' '-J' 'zuser@192.168.1.4' '-l' 'cisco' '10.1.21.10'
cisco@10.1.21.10's password:
[zrancid@jth-rocky8 ~]$ 'ssh' '-o' 'ConnectTimeout=10' '-J' 'zuser@192.168.1.4' '-l' 'cisco' '10.1.21.10'
Bad packet length 1349676916.
ssh_dispatch_run_fatal: Connection to UNKNOWN port 65535: message authentication code incorrect

In this case, a chain of two ssh worked fine. The commit adds support for
a sshchain via mode, as alternative to ssh which uses ProxyJump.